+++ lonessum/src/main/java/com/yapp/lonessum/domain/email/repository/EmailTokenRepository.java
@@ -1,6 +1,6 @@
-package com.yapp.lonessum.domain.user.repository;
+package com.yapp.lonessum.domain.email.repository;
 
-import com.yapp.lonessum.domain.user.entity.EmailTokenEntity;
+import com.yapp.lonessum.domain.email.entity.EmailTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmailTokenRepository extends JpaRepository<EmailTokenEntity, Long> {
